+++ src/vector2.rs
@@ -219,7 +219,7 @@ where
     }
 }
 
-impl Sum for Vector2<f64> {
+impl Sum for Vector2<f32> {
     fn sum<I>(iter: I) -> Self
     where
         I: Iterator<Item = Self>,
@@ -228,8 +228,8 @@ impl Sum for Vector2<f64> {
     }
 }
 
-impl From<(f64, f64)> for Vector2<f64> {
-    fn from(tuple: (f64, f64)) -> Self {
+impl From<(f32, f32)> for Vector2<f32> {
+    fn from(tuple: (f32, f32)) -> Self {
         Self { x: tuple.0, y: tuple.1 }
     }
 }
